1 var

+++ src/main/java/ru/kata/PP_3_1_3_Spring_Security_Bootstrap/controller/AdminController.java
@@ -71,6 +71,11 @@ public class AdminController {
 
     @DeleteMapping("/admin/{id}")
     public String deleteUser(@PathVariable("id") Long id) {
+        User user = userService.findUserById(id);
+        if (user == null) {
+            throw new RuntimeException("There is no user with ID = " +
+                    id + " in Database");
+        }
         userService.deleteUser(id);
         return "redirect:/admin";
     }
